add DB and liquibase for project, create table and change some field types because of it

+++ src/main/java/ru/stepanov/socksaccounting/repository/SocksRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.stepanov.socksaccounting.dao.SockDao;
 
 @Repository
-public interface SocksRepository extends JpaRepository<SockDao, Long> {
+public interface SocksRepository extends JpaRepository<SockDao, Integer> {
 }
